hives-africa-LMS/backend-codes/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database & Cache
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0

# Security & Authentication
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
bcrypt==4.0.1

# File Storage & Media
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# API Documentation
drf-spectacular==0.26.5

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.7.0
